requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
letterboxdpy==5.3.7
APScheduler==3.10.4
python-dotenv==1.0.0
pytz==2024.1
tmdbapis==1.2.11

# Testing dependencies
pytest==8.0.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0

# Linting and code quality
pylint==3.0.3
black==23.12.1
isort==5.13.2
pre-commit==3.6.0
